+++ dojo-lambda/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this project will be documented in this file.
 
+## [dojo-lambda-0.0.2]
+
+- Fix: deploy lambda
+
 ## [dojo-lambda-0.0.1]
 
 - Init module